refactor scorekeeper to separate files
